A couple of error corrections
I've corrected a couple of errors in this file so that it will compile. After you've created an instance of a class (e.g., myLosses is an instance of the Losses class) you will generally call methods using the instance variable (myLosses) rather than the class name (Losses). There are some exceptions to this but this is a good general rule to start with. Also, you had a call to a method (dislaynumWins) that did not exist in the Wins class. The method in that class is named "displayWins" so I changed the call to use that instead.